fix(archgate): inline MOBILE-004 scanner into rule file (v0.50.0 blocks local rule imports)
archgate v0.50.0's rule-file security scanner forbids rule files from importing
local modules (only node:path/url/util/crypto). The MOBILE-004 rule imported
../lint/desktopOnlyCommandGate → 'rule file blocked by security scanner' →
archgate fails on any PR touching ExocortexPlugin.ts (where MOBILE-004 fires),
including #3914. Inline the pure scanner functions into the rule file as named
exports (single source of truth), delete the helper, point the jest revert-verify
test at the rule file's exports — no drift. Unblocks #3914 + future command-touch PRs.
